Rename to object_detection_expression.py
I've also dropped AnnotationAccessor as there's no right file for it
now. It would be a circular import in annotation_expression.py.
