+++ lightly_studio/src/lightly_studio/core/dataset_query/__init__.py
@@ -1,7 +1,7 @@
-from .annotation_expression import ObjectDetectionField, ObjectDetectionQuery
 from .boolean_expression import AND, NOT, OR
 from .dataset_query import DatasetQuery
 from .image_sample_field import ImageSampleField
+from .object_detection_expression import ObjectDetectionField, ObjectDetectionQuery
 from .order_by import OrderByExpression, OrderByField
 from .video_sample_field import VideoSampleField
